src/Prototypes/engine/torch_impl/pyproject.toml
[project]
name = "torch-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "audiomentations>=0.42.0",
    "diskcache>=5.6.3",
    "hydra-core>=1.3.2",
    "librosa>=0.11.0",
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "pandas>=2.3.1",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch>=2.0.1",
    "torchaudio>=2.0.2",
    "torchvision>=0.15.2",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
]
